Addressed PR feedback
